requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain[openai]
langchain-huggingface
langchain-community
langchain-postgres
psycopg[binary,pool]
langgraph
atlassian-python-api
pydantic
python-dotenv
psycopg2-binary
lxml